requirements.txt
requests==2.29
coloredlogs==15.0
colorama==0.4
python-dotenv==1.0.0
tabulate==0.9
